server/requirements.txt
flask
flask-cors
transformers
faiss-cpu
markdown
flask-socketio
PyPDF2
torch
torchvision
torchaudio
